% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/input_argument.R
\name{input_argument}
\alias{input_argument}
\title{An input function}
\usage{
input_argument(input = TRUE)
}
\arguments{
\item{input}{}
}
\description{
Checking for input argument value being passed in function input_argument
}
\examples{
input_function()
}
\keyword{mlasiapackage}

